Move the UI code to ui.go
